email-widget/requirements.txt
html2text==2020.1.16
python-dotenv==1.0.0
